Fix fprintf conversion specifiers
A run of cppcheck against the code showed a number of type mismatches
between what the code provided vs what fprintf attempted to print.
This could result in some information lost or confused.

Reviewed-on: https://gerrit.openafs.org/13135
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/vol/xfs_size_check.c
@@ -126,9 +126,9 @@ CheckPartitions()
 					   nAvail * sizeof(partInfo));
 		    if (!partList) {
 			printf
-			    ("Failed to %salloc %d bytes for partition list.\n",
+			    ("Failed to %salloc %u bytes for partition list.\n",
 			     (nAvail == ALLOC_STEP) ? "m" : "re",
-			     nAvail * sizeof(partInfo));
+			     (unsigned int) (nAvail * sizeof(partInfo)));
 			exit(1);
 		    }
 		}
